Use pattern regex in schema to establish date format
We don't currently support a date format that is not a datetime, and we
decided in today's team 1 meeting that including a time in this field
may require extra effort to ensure that we aren't tripped up by timezone
localization. For now, we want to use pattern matching, but may
establish a more formal date format in the future.