CIRCSTORE-183: update RMB from 26.2.3 to 29.2.x changes due to review
Due to a misunderstanding on my part, some indexes were unnecessarily
removed from schema.json.

Checking for the error type `RTFConsts.VALIDATION_FIELD_ERROR` is not
necessary. There is only one error type to be expected.

Remove commented out test code.

Explicitly include HttpStatus codes as static imports to reduce line
lengths.

Rename the WebClient test class into OkapiHttpClient
(This class used to be called HttpClient).

As suggested by @bohdan-suprun, use a valid foreign key instead of
"1234567890". A separate test needs to be made for testing invalid
keys. This test is not for invalid UUID's but instead for
non-existent foreign keys.

With the HTTP 500 worked-around by not using an invalid UUID, the
status handling can be reduced to:
1) Handling the upsert using the previous approach.
2) Handling the new behavior via processing the Errors class, but only
   analyzing the client errors (HTTP 4xx status codes only)

A new private method is added (where appropriate) to be used for the
On Upsert only cases. There is a comment associated with it to
inform any developers that this method should be removed when the
deprecated MyPgUtil.putUpsert204() method is dealt with.

It turns out that the `RequestPreferencesAPI.java` did not need as
many checks as some of the other classes. Remove the additional checks.

Implement a custom matcher for Okapi Response Status codes to cleanup
the code. There is already an HttpResponseStatusCodeMatchers, but that
has slightly different design. Ideally the
OkapiResponseStatusCodeMatchers will eventually completely replace
HttpResponseStatusCodeMatchers.

Minor syntax or code structural changes.

Do not remove the lines, change "tOps": "ADD" to "tOps": "DELETE"
so that the index gets removed.

Ensure that itemId and userId exist as b-tree indexes.

Implement OkapiResponseUtil Utility class.