CIRCSTORE-222: Replace LoanPolicy Drools .drl rules by plain Java code
RMB v30.0.1 ( https://issues.folio.org/browse/RMB-525 ) removes
RestVerticle's automatic Drools rules validation because Drools
is a big library and mod-circulation-storage is the only module
that uses it.

It is possible to put the drools code that was deleted in RMB into
mod-circulation-storage: https://github.com/folio-org/raml-module-builder/pull/672/files
However, the rules are very simple:
https://github.com/folio-org/mod-circulation-storage/tree/v12.0.0/src/main/resources/rules
Using Drools is over-engineering because a few Java if-clauses
suffice and converting the Drools rules to Java is trivial.