CIRCSTORE-263: Broken logging for mod-circulation-storage
I went through the document linked in the description of this issue
and applied all the noted changes.  Most of them had already been done,
i just added a few things that seem to have been missed.  Despite the
name of this issue, the fixes for logging seem to be in version 32,
so I'm not sure this will actually impact any problems this module
is having with logging.

I also had a test failing on build.  It looked to me like the date
generation for the two loans in this test were a bit screwy-the date
labeled "tomorrow" was actually today.  I tweaked this test a bit
so that one loan was indeed set with a date one day from the present
and the other a date one day in the past, then set it to look for
anything with a date greater than the present day.  That seems to
work.

refs: https://issues.folio.org/browse/CIRCSTORE-263