CIRCSTORE-391: /scheduled-request-expiration hangs on invalid URL or unknown host
Code in RequestExpiryImpl: onComplete should be called after compose, but
it is called within compose. If getTlrSettings() returns a failed Future
no 500 response is generated and the request hangs forever.

In addition getTlrSettings() may throw an exception that is not caught,
this also makes the request hang forever. The exception origins from OkapiClient.